+++ ios/flutter_native_contact_picker/Package.swift
@@ -1,0 +1,28 @@
+// swift-tools-version: 5.9
+// The swift-tools-version declares the minimum version of Swift required to build this package.
+
+import PackageDescription
+
+let package = Package(
+    name: "flutter_native_contact_picker",
+    platforms: [
+        .iOS("13.0"),
+    ],
+    products: [
+        .library(name: "flutter-native-contact-picker", targets: ["flutter_native_contact_picker"]),
+    ],
+    dependencies: [
+        .package(name: "FlutterFramework", path: "../FlutterFramework"),
+    ],
+    targets: [
+        .target(
+            name: "flutter_native_contact_picker",
+            dependencies: [
+                .product(name: "FlutterFramework", package: "FlutterFramework"),
+            ],
+            resources: [
+                .process("PrivacyInfo.xcprivacy"),
+            ]
+        ),
+    ]
+)