+++ ios/flutter_native_contact_picker/Sources/flutter_native_contact_picker/FlutterNativeContactPickerPlugin.swift
@@ -106,4 +106,3 @@ class PhoneNumberPickerHandler: PickerHandler {
         }
     }
 }
-